refactor: 컴포넌트 구성 파일들을 컴포넌트 단위의 폴더로 묶어 재정리
