+++ src/components/App/reset.css
@@ -58,6 +58,18 @@ textarea {
   -webkit-user-select: auto;
 }
 
+/* Chrome, Safari, Edge, Opera - removes spin buttons in input[type='number'] */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+
+/* Firefox - removes spin buttons in input[type='number'] */
+input[type='number'] {
+  -moz-appearance: textfield;
+}
+
 /* revert the 'white-space' property for textarea elements on Safari */
 textarea {
   white-space: revert;
