+++ src/components/CardExpirationInput/CardExpirationInput.stories.tsx
@@ -1,5 +1,4 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { fn } from '@storybook/test';
 import CardExpirationInput from './CardExpirationInput';
 
 const meta = {
@@ -13,8 +12,7 @@ type Story = StoryObj<typeof CardExpirationInput>;
 
 export const Default: Story = {
   args: {
-    isMonthValid: { isValid: true, errorMessage: '' },
-    isYearValid: { isValid: true, errorMessage: '' },
-    onChangeExpireDate: fn(),
+    setMonth: () => {},
+    setYear: () => {},
   },
 };
