+++ src/components/CardNumberInput/CardNumberInput.stories.tsx
@@ -1,5 +1,4 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { fn } from '@storybook/test';
 import CardNumberInput from './CardNumberInput';
 
 const meta = {
@@ -13,7 +12,6 @@ type Story = StoryObj<typeof CardNumberInput>;
 
 export const Default: Story = {
   args: {
-    isCardNumbersValid: { validStates: [true, true, true, true], errorMessage: '' },
-    onChangeCardNumbers: fn(),
+    setCardNumber: () => {},
   },
 };
