+++ src/components/CardOwnerInput/CardOwnerInput.stories.tsx
@@ -1,5 +1,4 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { fn } from '@storybook/test';
 import CardOwnerInput from './CardOwnerInput';
 
 const meta = {
@@ -13,7 +12,6 @@ type Story = StoryObj<typeof CardOwnerInput>;
 
 export const Default: Story = {
   args: {
-    isOwnerValid: { isValid: true, errorMessage: '' },
-    onChangeOwner: fn(),
+    setOwner: () => {},
   },
 };
