+++ src/components/CardPreview/CardPreview.style.tsx
@@ -5,8 +5,8 @@ export const Card = styled.div`
   height: 132px;
   border-radius: 4px;
   padding: 8px 12px;
-  background: var(--grey-400);
-  box-shadow: 3px 3px 5px 0px var(--card-shadow);
+  background: #333333;
+  box-shadow: 3px 3px 5px 0px #00000040;
 `;
 
 export const CardHeader = styled.header`
@@ -18,8 +18,8 @@ export const ChipBox = styled.div`
   width: 36px;
   height: 22px;
   border-radius: 4px;
-  background: var(--card-chip);
-  border: 0.5px solid var(--card-chip-border);
+  background: #ddcd78;
+  border: 0.5px solid #ddcd781a;
 `;
 
 export const LogoBox = styled.div`
@@ -41,7 +41,7 @@ export const InfoBox = styled.p<{ $length?: number }>`
   width: ${(props) => props.$length && `calc(100% / ${props.$length})`};
   height: 20px;
   color: white;
-  font-size: var(--font-size-xl);
+  font-size: 14px;
   font-weight: 500;
   line-height: 20px;
   letter-spacing: 0.08em;
