--- src/components/Input.tsx
@@ -1,12 +1,0 @@
-import { InputHTMLAttributes } from 'react';
-import * as S from '../styles/Input.style';
-
-interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
-  isValid: boolean;
-}
-
-function Input({ isValid, ...rest }: InputProps) {
-  return <S.Input $isValid={isValid} {...rest} />;
-}
-
-export default Input;
