+++ src/components/common/TitleContainer/TitleContainer.style.tsx
@@ -8,14 +8,14 @@ export const TitleContainer = styled.div`
 `;
 
 export const Title = styled.h2`
-  font-size: var(--font-size-h2);
+  font-size: 18px;
   font-weight: 700;
   line-height: 26px;
 `;
 
 export const SubTitle = styled.p`
-  color: var(--grey-300);
-  font-size: var(--font-size-sm);
+  color: #8b95a1;
+  font-size: 10px;
   font-weight: 400;
   line-height: 13px;
 `;
