--- src/index.css
@@ -1,11 +1,0 @@
-/* Chrome, Safari, Edge, Opera */
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-/* Firefox  */
-input[type='number'] {
-  -moz-appearance: textfield;
-}
