--- src/stories/CardNumberInput.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import CardNumberInput from '../components/CardNumberInput';
-
-const meta = {
-  title: 'CardNumberInput',
-  component: CardNumberInput,
-} satisfies Meta<typeof CardNumberInput>;
-
-export default meta;
-
-type Story = StoryObj<typeof CardNumberInput>;
-
-export const Default: Story = {
-  args: {
-    setCardNumber: () => {},
-  },
-};
