--- src/stories/CardOwnerInput.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import CardOwnerInput from '../components/CardOwnerInput';
-
-const meta = {
-  title: 'CardOwnerInput',
-  component: CardOwnerInput,
-} satisfies Meta<typeof CardOwnerInput>;
-
-export default meta;
-
-type Story = StoryObj<typeof CardOwnerInput>;
-
-export const Default: Story = {
-  args: {
-    setOwner: () => {},
-  },
-};
