--- src/styles/App.style.tsx
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-
-export const AppLayout = styled.main`
-  padding: 77px 30px;
-`;
-
-export const CardPreviewBox = styled.section`
-  display: flex;
-  justify-content: center;
-  width: 100%;
-`;
-
-export const CardForm = styled.section`
-  display: flex;
-  flex-direction: column;
-  row-gap: 16px;
-  margin-top: 45px;
-`;
