--- src/styles/TitleContainer.style.tsx
@@ -1,21 +1,0 @@
-import styled from 'styled-components';
-
-export const TitleContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  row-gap: 3px;
-  margin-bottom: 16px;
-`;
-
-export const Title = styled.h2`
-  font-size: 18px;
-  font-weight: 700;
-  line-height: 26px;
-`;
-
-export const SubTitle = styled.p`
-  color: #8b95a1;
-  font-size: 10px;
-  font-weight: 400;
-  line-height: 13px;
-`;
